+++ src/repository.rs
@@ -1187,7 +1187,7 @@ pub trait ProcessCa: Sized + Send + Sync {
 
 //------------ Helper Functions ----------------------------------------------
 
-#[allow(clippy::manual_strip)] // str::strip_prefix not in 1.42
+#[allow(clippy::manual_strip)] // str::strip_prefix not in 1.43
 fn uri_relative_to<'a>(
     uri: &'a uri::Rsync,
     other: &uri::Rsync