+++ build.rs
@@ -3,9 +3,9 @@ use rustc_version::{Version, version};
 
 fn main() {
     let version = version().expect("Failed to get rustc version.");
-    if version < Version::parse("1.43.0").unwrap() {
+    if version < Version::parse("1.44.0").unwrap() {
         eprintln!(
-            "\n\nAt least Rust version 1.43 is required.\n\
+            "\n\nAt least Rust version 1.44 is required.\n\
              Version {} is used for building.\n\
              Build aborted.\n\n",
              version);