+++ Changelog.md
@@ -7,6 +7,9 @@ Breaking Changes
 
 New
 
+* Status information is now available in JSON format at `/api/v1/status`.
+  ([#437])
+
 Bug Fixes
 
 Other Changes
@@ -14,6 +17,7 @@ Other Changes
 * Upgrade [rpki-rs] to 0.11 and drop now unnecessary separate dependency
   to [rpki-rtr]. ([#443])
 
+[#437]: https://github.com/NLnetLabs/routinator/pull/437
 [#443]: https://github.com/NLnetLabs/routinator/pull/443
 [rpki-rs]: https://github.com/NLnetLabs/rpki-rs/
 [rpki-rtr]: https://github.com/NLnetLabs/rpki-rtr/