+++ Changelog.md
@@ -13,8 +13,11 @@ Other Changes
 
 * Upgrade [rpki-rs] to 0.11 and drop now unnecessary separate dependency
   to [rpki-rtr]. ([#443])
+* Upgrade Tokio-related dependencies to new version based on Tokio 1.0.
+  ([#444])
 
 [#443]: https://github.com/NLnetLabs/routinator/pull/443
+[#444]: https://github.com/NLnetLabs/routinator/pull/444
 [rpki-rs]: https://github.com/NLnetLabs/rpki-rs/
 [rpki-rtr]: https://github.com/NLnetLabs/rpki-rtr/